DCS-1348 restructuring pages to match routes
