+++ integration_tests/integration/bookedtoday/arrivals/choosePrisoner.spec.ts
@@ -1,6 +1,6 @@
-import ChoosePrisonerPage from '../../../pages/choosePrisoner'
-import ConfirmArrivalPage from '../../../pages/confirmArrival'
-import CheckCourtReturnPage from '../../../pages/checkCourtReturn'
+import ChoosePrisonerPage from '../../../pages/bookedtoday/choosePrisoner'
+import ConfirmArrivalPage from '../../../pages/bookedtoday/arrivals/confirmArrival'
+import CheckCourtReturnPage from '../../../pages/bookedtoday/arrivals/courtreturns/checkCourtReturn'
 import Page from '../../../pages/page'
 import Role from '../../../../server/authentication/role'
 import expectedArrivals from '../../../mockApis/responses/expectedArrivals'
