+++ integration_tests/integration/temporaryabsences/temporaryAbsences.spec.ts
@@ -1,7 +1,7 @@
 import Page from '../../pages/page'
 import Role from '../../../server/authentication/role'
-import TemporaryAbsencesPage from '../../pages/temporaryAbsences'
-import CheckTemporaryAbsencePage from '../../pages/checkTemporaryAbsence'
+import TemporaryAbsencesPage from '../../pages/temporaryabsences/temporaryAbsences'
+import CheckTemporaryAbsencePage from '../../pages/temporaryabsences/checkTemporaryAbsence'
 import temporaryAbsences from '../../mockApis/responses/temporaryAbsences'
 
 context('A user can view all current temporary absences', () => {
