+++ integration_tests/pages/bookedtoday/arrivals/checkAnswers.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
 export default class CheckAnswersPage extends Page {
   constructor() {
