+++ integration_tests/pages/bookedtoday/arrivals/confirmAddedToRoll.ts
@@ -1,14 +1,13 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
-export default class ConfirmTransferAddedToRollPage extends Page {
+export default class ConfrimAddedToRollPage extends Page {
   constructor() {
     super('has been added to the establishment roll')
   }
 
-  static goTo(id: string): ConfirmTransferAddedToRollPage {
-    cy.visit(`/prisoners/${id}/confirm-transfer`)
-    return Page.verifyOnPage(ConfirmTransferAddedToRollPage)
-  }
+  confirmationBanner = (): PageElement => cy.get(`[data-qa=confirmation-banner]`)
+
+  confirmationParagraph = (): PageElement => cy.get(`[data-qa=confirmation-paragraph]`)
 
   addAnotherToRoll = (): PageElement => cy.get(`[data-qa=add-another-to-roll]`)
