+++ integration_tests/pages/bookedtoday/arrivals/confirmArrival.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
 export default class ConfirmArrivalPage extends Page {
   constructor() {
