+++ integration_tests/pages/bookedtoday/arrivals/courtreturns/checkCourtReturn.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../../page'
 
 export default class CheckCourtReturnPage extends Page {
   constructor() {
