+++ integration_tests/pages/bookedtoday/arrivals/courtreturns/confirmCourtReturnAddedToRoll.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../../page'
 
 export default class ConfirmCourtReturnAddedToRollPage extends Page {
   constructor() {
