+++ integration_tests/pages/bookedtoday/arrivals/imprisonmentStatus.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
 export default class ImprisonmentStatusPage extends Page {
   constructor() {
