+++ integration_tests/pages/bookedtoday/arrivals/movementReasons.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
 export default class MovementReasonsPage extends Page {
   constructor() {
