+++ integration_tests/pages/bookedtoday/choosePrisoner.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../page'
 
 export default class ChoosePrisonerPage extends Page {
   constructor() {
