+++ integration_tests/pages/bookedtoday/transfers/checkTransfer.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../../page'
 
 export default class CheckTransferPage extends Page {
   constructor() {
