+++ integration_tests/pages/temporaryabsences/confirmTemporaryAbsenceAddedToRoll.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../page'
 
 export default class ConfirmTemporaryAbsenceAddedToRollPage extends Page {
   constructor() {
