+++ integration_tests/pages/temporaryabsences/temporaryAbsences.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../page'
 
 export default class TemporaryAbsencesPage extends Page {
   constructor() {
